# src/lib/api/generated/ruledOffAPI.schemas.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RuledOff API
 * OpenAPI spec version: 1.0
 */
export interface ExchangeMagicLinkDto {
  /** The opaque token from the emailed magic link. */
  token: string;
}

export interface ClientResponse {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  notes: string | null;
  /**
   * Day of the month the books are due to close.
   * @nullable
   */
  close_day: number | null;
  /** Start next month's close from the template automatically. */
  auto_chase: boolean;
  /** @nullable */
  qbo_realm_id: string | null;
  created_at: string;
}

export interface CreateClientDto {
  name: string;
  email: string;
  phone?: string;
  notes?: string;
  /**
   * @minimum 1
   * @maximum 28
   */
  closeDay?: number;
}

export interface UpdateClientDto {
  name?: string;
  email?: string;
  phone?: string;
  notes?: string;
  /**
   * @minimum 1
   * @maximum 28
   */
  closeDay?: number;
}

export interface DeletedResponse {
  deleted: boolean;
}

export interface AttachmentResponse {
  path: string;
  name: string;
  size: number;
  mime: string;
  uploaded_at: string;
}

export type ItemResponseState = typeof ItemResponseState[keyof typeof ItemResponseState];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ItemResponseState = {
  requested: 'requested',
  nudged: 'nudged',
  answered: 'answered',
  accepted: 'accepted',
} as const;

export interface ItemResponse {
  id: string;
  type: string;
  title: string;
  state: ItemResponseState;
  /** Derived: the item is accepted. */
  ruledOff: boolean;
  /** Derived: still open past the grace period. */
  overdue: boolean;
  daysOpen: number;
  /** @nullable */
  answerText: string | null;
  attachments: AttachmentResponse[];
  createdAt: string;
  /** @nullable */
  answeredAt: string | null;
  /** @nullable */
  acceptedAt: string | null;
}

export type CreateItemDtoType = typeof CreateItemDtoType[keyof typeof CreateItemDtoType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateItemDtoType = {
  transaction: 'transaction',
  document: 'document',
  questionnaire: 'questionnaire',
} as const;

export interface CreateItemDto {
  type: CreateItemDtoType;
  title: string;
  note?: string;
}

export interface SignedUrlResponse {
  url: string;
}

export interface SignUploadDto {
  filename: string;
  mime: string;
  /** Bytes. Rejected above the bucket limit. */
  size: number;
}

export interface SignedUploadResponse {
  /** Server-derived object path. Never caller supplied. */
  path: string;
  token: string;
  signedUrl: string;
  bucket: string;
}

export interface AnswerItemDto {
  /** What the client typed back */
  answerText: string;
}

export interface ConfirmUploadDto {
  /** Storage path returned by the sign step */
  path: string;
  name: string;
  size: number;
  mime: string;
}

export interface PauseDto {
  /** True to hold the chase, false to resume it. */
  paused: boolean;
}

export interface PauseResponse {
  clientId: string;
  paused: boolean;
}

export interface ReminderEventResponse {
  type: string;
  occurred_at: string;
}

export interface ReminderResponse {
  id: string;
  /**
   * @minimum 1
   * @maximum 4
   */
  level: number;
  /** friendly, specific-deadline, consequence or weekly */
  levelLabel: string;
  channel: string;
  /** @nullable */
  sentAt: string | null;
  /** @nullable */
  stoppedReason: string | null;
  createdAt: string;
  events: ReminderEventResponse[];
}

export interface EntitlementFeatures {
  forecast: boolean;
  aiAnalyst: boolean;
  responsiveness: boolean;
  closeReceipt: boolean;
  smsEscalation: boolean;
  bulkChase: boolean;
  autoChase: boolean;
  emailPreview: boolean;
  autoEscalatingReminders: boolean;
}

export type EntitlementsResponseTier = typeof EntitlementsResponseTier[keyof typeof EntitlementsResponseTier];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EntitlementsResponseTier = {
  free: 'free',
  pro: 'pro',
  scale: 'scale',
} as const;

export type EntitlementsResponseStatus = typeof EntitlementsResponseStatus[keyof typeof EntitlementsResponseStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EntitlementsResponseStatus = {
  trialing: 'trialing',
  active: 'active',
  past_due: 'past_due',
  paused: 'paused',
  canceled: 'canceled',
  expired: 'expired',
} as const;

export interface EntitlementsResponse {
  tier: EntitlementsResponseTier;
  status: EntitlementsResponseStatus;
  /** @nullable */
  trialEndsAt: string | null;
  /** @nullable */
  currentPeriodEnd: string | null;
  features: EntitlementFeatures;
}